procedures for msg4

+++ openair1/PHY/LTE_TRANSPORT/proto_NB_IoT.h
@@ -124,7 +124,7 @@ void dlsch_sib_scrambling_rar_NB_IoT(LTE_DL_FRAME_PARMS     *frame_parms,
                                   int                    tot_bits,                // total number of bits to transmit
                                   uint16_t                Nf,              // Nf is the frame number (0..9)
                                   uint8_t                Ns,
-                                  uint16_t               rnti);  
+                                  uint32_t               rnti);  
 /*
 int scrambling_npbch_REs_rel_14(LTE_DL_FRAME_PARMS      *frame_parms,
                                 int32_t                 **txdataF,
